clients: move Section and InvoicePreview to Shared
